requirements.txt
pillow==11.2.1
scratchattach==2.1.13